Refactor Dashboard components and create new modular components
- Extracted DashboardHeader, DashboardWelcome, DashboardStatsCards, CreateShortLinkCard, LinksSection, and UrlRow into separate components for better organization and maintainability.
- Updated Dashboard component to utilize new modular components, improving readability and structure.
- Removed unused imports and code, streamlining the Dashboard component.
- Enhanced the CreateShortLinkCard component to handle URL input and submission.
- Improved LinksSection to manage link display and search functionality.
- Updated UrlRow to handle individual link actions and display metadata.
- Removed hardcoded Supabase anon key in redirect.js for better security practices.
